Fix warnings on x64 build: casting pointer to int, size_t related etc.

+++ Library_MMDAgent/include/InfoText.h
@@ -29,8 +29,8 @@ private:
    struct LINK {
       char *url;
       float x1, y1, x2, y2;
-      int textloc;
-      int textlen;
+      size_t textloc;
+      size_t textlen;
       struct LINK *next;
       LINK() {
          url = NULL;
