+++ Library_MMDFiles/src/lib/VMD.cpp
@@ -123,7 +123,7 @@ void VMD::addBoneMotion(const char *name)
    link->next = m_boneLink;
    m_boneLink = link;
 
-   m_name2bone.add(name, strlen(name), bmNew);
+   m_name2bone.add(name, (int)strlen(name), bmNew);
 }
 
 /* VMD::addFaceMotion: add new face motion to list */
@@ -142,7 +142,7 @@ void VMD::addFaceMotion(const char *name)
 
    link->next = m_faceLink;
    m_faceLink = link;
-   m_name2face.add(name, strlen(name), fmNew);
+   m_name2face.add(name, (int)strlen(name), fmNew);
 }
 
 /* VMD::getBoneMotion: find bone motion by name */
@@ -153,7 +153,7 @@ BoneMotion* VMD::getBoneMotion(const char *name)
    if (name == NULL)
       return NULL;
 
-   if (m_name2bone.search(name, strlen(name), (void **)&bm) == true)
+   if (m_name2bone.search(name, (int)strlen(name), (void **)&bm) == true)
       return bm;
 
    return NULL;
@@ -167,7 +167,7 @@ FaceMotion* VMD::getFaceMotion(const char *name)
    if(name == NULL)
       return NULL;
 
-   if (m_name2face.search(name, strlen(name), (void **)&fm) == true)
+   if (m_name2face.search(name, (int)strlen(name), (void **)&fm) == true)
       return fm;
 
    return NULL;
